Created the Analytic class to store/retrieve (setters and getters) the stats #25
The Analytic class is able to 1. count the number of outgoing messages, 2. count the number of incoming messages, 3. save the person's name to whom a message was last sent to, 4. save the person's name a message was last received from. Also, the runSample.java is class where the Analytic class is tested